# BackEnd/ADMIN_DA3/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get logged-in user details",
                "operationId": "6d010834430854e4df1a215e5771aa96",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/baiviet/search": {
            "post": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "f4b6a4fc9e7ea6c2f475c2a8d625d87e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "totalPage",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/api/baiviet/get/{id}": {
            "get": {
                "tags": [
                    "BaiViet"
                ],
                "summary": "Get logged-in user details",
                "operationId": "ffbe4e3a71236593db3eb3600f138acb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của bài viết",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/binhluan/search": {
            "post": {
                "tags": [
                    "BinhLuan"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "db54a3afb8ea91cd8071903bbfd2c298",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "totalPage",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/binhluan/": {
            "get": {
                "tags": [
                    "BinhLuan"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "bcce5281a6c2e152fc5c77ac941ea34b",
                "parameters": [
                    {
                        "name": "total",
                        "in": "path",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/binhluan/delete/{id}": {
            "delete": {
                "tags": [
                    "BinhLuan"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "9863149372066bceec1519c4f7fe3457",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/binhluan/deletes": {
            "delete": {
                "tags": [
                    "BinhLuan"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "f6f0abe874fa8190255205072f6e4a3f",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/binhluan/save/{id}": {
            "get": {
                "tags": [
                    "BinhLuan"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "97e7bada5483f94d3d8f9a75c44f3849",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "MaBaiViet",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "MaModel",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "MaTaiKhoan",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "NoiDung",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/binhluan/get/{id}": {
            "get": {
                "tags": [
                    "BinhLuan"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "8ba322fa6e413f1d00e6d1ad9b65bea9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}